+++ modules/genealogy-people-filament/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added the person attributes editor to the Filament resource.
+- Added a Filament life-status transition action.
 
 ## 1.0.0